+++ src/include/test/unit_tdb_partitioned_matrix.cc
@@ -253,7 +253,7 @@ TEST_CASE("test different combinations", "[tdb_partitioned_matrix]") {
   std::vector<uint32_t> dimensions_options = {3, 77};
   std::vector<int> num_parts_options = {1, 3, 99, 989, 2143};
   for (auto num_vectors : num_vectors_options) {
-    for (uint32_t dimensions : dimensions_options) {
+    for (uint64_t dimensions : dimensions_options) {
       for (auto num_parts : num_parts_options) {
         auto training_set =
             ColMajorMatrix<feature_type>(dimensions, num_vectors);
@@ -337,7 +337,7 @@ TEST_CASE(
       (std::filesystem::temp_directory_path() / "ids").string();
 
   size_t num_vectors = 10000;
-  uint32_t dimensions = 128;
+  uint64_t dimensions = 128;
 
   // Setup data.
   {
